LayoutTests/platform/chromium/http/tests/security/contentSecurityPolicy/eval-blocked-in-about-blank-iframe-expected.txt

ALERT: /PASS/
CONSOLE MESSAGE: line 1: Uncaught Error: Code generation from strings disallowed for this context
 Eval should be blocked in the iframe, but inline script should be allowed.
